Add simple frontend for builds api - Refactor publish (projects) app into separate modules
