+++ src/Publish/constants.ts
@@ -1,0 +1,46 @@
+const newTechEmail =
+  "mailto:hank@compute.studio?subject=Requesting%20New%20App%20Technology&body=I%20am%20interested%20in%20publishing%20an%20app%20using:";
+
+const techLinks = {
+  "python-paramtools": "https://paramtools.dev",
+  bokeh: "https://bokeh.org",
+  dash: "https://dash.plotly.com/",
+  streamlit: "https://streamlit.io",
+};
+
+const techDocsLinks = {
+  "python-paramtools": "https://docs.compute.studio/publish/guide/",
+  bokeh: "https://bokeh.org",
+  dash: "https://dash.plotly.com/",
+  streamlit: "https://streamlit.io",
+};
+
+const techGuideLinks = {
+  bokeh: "https://docs.compute.studio/publish/data-viz/guide.html#bokeh",
+  dash: "https://docs.compute.studio/publish/data-viz/guide.html#dash",
+  "python-paramtools": "https://docs.compute.studio/publish/model/guide.html",
+  streamlit: "https://docs.compute.studio/publish/data-viz/guide.html#streamlit",
+};
+
+const techTitles = {
+  dash: "Dash",
+  bokeh: "Bokeh",
+  "python-paramtools": "ParamTools",
+  streamlit: "Streamlit",
+};
+
+const inputStyle = {
+  width: "100%",
+};
+
+const requiredMessage = "This field is required.";
+
+export {
+  newTechEmail,
+  techLinks,
+  techDocsLinks,
+  techGuideLinks,
+  techTitles,
+  inputStyle,
+  requiredMessage,
+};
