+++ src/Publish/pages/index.tsx
@@ -1,0 +1,6 @@
+import { NewProject } from "./NewProject";
+import { ProjectDetail } from "./Project";
+import { BuildHistory } from "./BuildHistory";
+import { BuildPage } from "./Build";
+
+export { NewProject, ProjectDetail, BuildHistory, BuildPage };
